docs: Removed prune and prune_func constructor parameters.
Pruning is not a suitable concern of the Node's constructor. It should be handled by the tree constructor. This is because pruning should happen to the tree after it is completely built. It's still okay to have the prune function however, because this is what the tree class will call.

See GitHub Issue #2.