src: Implemented Node.num_positive/num_negative functions.
The num_positive function is trivial, it simply returns the number of data points with the positive class value. However, the num_negative function has to find the sum of all class supports excluding the positive class. This is why the num_positive function was included (so that the number of positive/negative are found in the same way).

See GitHub Issue #2.